+++ src/plugins/notesSearcher/index.tsx
@@ -18,7 +18,7 @@ import settings from "./settings";
 
 export default definePlugin({
     name: "NotesSearcher",
-    description: "Allows you to open modal with all of your notes and search throught them by UserID, Note text and Global/Username if user is cached",
+    description: "Allows you to open a modal with all of your notes and search through them by user ID, note text, and username",
     authors: [Devs.Vishnya],
     settings,
     patches: [
